Add loader thread for torrent restore
This change allow main thread to process events while torrents are read
from disk. This fixes long UI freeze when starting with a large amount
of torrents to restore. With few thousand it can take few to even
dozens of seconds to load.

Also this change nicely interleaves loading torrent with processing
alerts of libtorrent.
